requirements/uv-repro.txt
# Reproducible pip/uv environment approximating docker/environment.yml (conda).
# Notes:
# - Conda-only runtime libs (e.g., cudatoolkit, ffmpeg, libgcc, etc.) are intentionally omitted.
# - Install the package itself with: uv pip install -e .
# - Install this file with: uv pip install -r requirements/uv-repro.txt
#
# Extra research utilities (installed from source via git)
# - eff-physics-learn-dataset: dataset utilities package (uv/pyproject-based)
# - metrics-structures: shared metrics data structures
eff-physics-learn-dataset @ git+https://github.com/yoelt11/eff-physics-learn-dataset.git
metrics-structures @ git+https://github.com/yoelt11/metrics-structures

# PyTorch family (keep versions aligned with torchvision/torchaudio)
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
torchdata>=0.11.0

# Core scientific stack (from conda section)
numpy>=1.24.0
pandas==1.4.4
xarray==0.20.1
h5py>=3.10.0
zarr==2.8.1
numcodecs>=0.12.0
numexpr==2.8.3
bottleneck==1.3.5
pillow==9.2.0
psutil==5.9.0
requests>=2.32.2
bokeh==2.4.3
dask==2022.7.0
distributed==2022.7.0

# From pip section in docker/environment.yml
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
docstring-parser==0.15
fonttools==4.37.4
frozenlist==1.3.1
fsspec>=2023.5.0
gcsfs>=2023.5.0
google-api-core==2.10.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
hydra-core==1.2.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
jsonargparse[signatures]>=4.27.7
kiwisolver==1.4.4
markdown==3.4.1
# eff-physics-learn-dataset requires matplotlib>=3.8.0
matplotlib>=3.8.0
multidict==6.0.2
oauthlib==3.2.1
omegaconf==2.2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeprecate==0.3.2
pygments==2.13.0
pytorch-lightning==1.7.7
pyyaml==6.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torchmetrics==0.10.0
tqdm>=4.66.0
werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
netcdf4==1.6.4
h5netcdf==1.1.0

